requirements.txt
ultralytics
opencv-python
streamlit
pandas
win10toast
pillow
numpy
plotly
psycopg2-binary
python-dotenv
SQLAlchemy
fastapi
uvicorn
pydantic
mediapipe
websockets
onnxruntime
requests
